feat: aggregation decorator
refactor the aggregation into a bucketList
