--- src/monitor/index.test.ts
@@ -1,10 +1,0 @@
-import tap from "tap";
-import myfunction from "./index";
-
-tap.test("Describe block", (t) => {
-  t.plan(1);
-  t.test("basics", (t) => {
-    t.equal(myfunction("abc"), "abc");
-    t.end();
-  });
-});
