--- src/monitor/index.ts
@@ -1,3 +1,0 @@
-const myfunction = (text: string): string => text;
-
-export default myfunction;
